move addInternalCheckpointingEvent to funnel utils
