+++ packages/engine/paima-funnel/src/funnels/avail/funnel.ts
@@ -26,7 +26,11 @@ import {
   slotToTimestamp,
   timestampToSlot,
 } from './utils.js';
-import { buildParallelBlockMappings, findBlockByTimestamp } from '../../utils.js';
+import {
+  addInternalCheckpointingEvent,
+  buildParallelBlockMappings,
+  findBlockByTimestamp,
+} from '../../utils.js';
 
 type BlockData = {
   number: number;
@@ -534,50 +538,6 @@ export function composeChainData(baseChainData: ChainData[], cdeData: BlockData[
 
     return blockData;
   });
-}
-
-// This adds the internal event that updates the last block point. This is
-// mostly to avoid having to do a binary search each time we boot the
-// engine. Since we need to know from where to start searching for blocks in
-// the timestamp range.
-function addInternalCheckpointingEvent(
-  chainData: ChainData[],
-  mapBlockNumber: (mainchainNumber: number) => number,
-  chainName: string,
-  // FIXME: not really clear why this ignore is needed
-  // eslint-disable-next-line @typescript-eslint/no-redundant-type-constituents
-  eventType: InternalEventType.EvmLastBlock | InternalEventType.AvailLastBlock
-): void {
-  for (const data of chainData) {
-    const originalBlockNumber = mapBlockNumber(data.blockNumber);
-    // it's technically possible for this to be null, because there may not be
-    // a block of the sidechain in between a particular pair of blocks or the
-    // original chain.
-    //
-    // in this case it could be more optimal to set the block number here to
-    // the one in the next block, but it shouldn't make much of a difference.
-    if (!originalBlockNumber) {
-      continue;
-    }
-
-    if (!data.internalEvents) {
-      data.internalEvents = [];
-    }
-    data.internalEvents.push({
-      type: eventType,
-      // this is the block number in the original chain, so that we can resume
-      // from that point later.
-      //
-      // there can be more than one block here, for example, if the main
-      // chain produces a block every 10 seconds, and the parallel chain
-      // generates a block every second, then there can be 10 blocks.
-      // The block here will be the last in the range. Losing the
-      // information doesn't matter because there is a transaction per main
-      // chain block, so the result would be the same.
-      block: originalBlockNumber,
-      network: chainName,
-    });
-  }
 }
 
 async function readFromBaseFunnel(
